+++ crates/objc-sys/extern/exception.m
@@ -9,7 +9,7 @@ id objc_retain(id value);
 /// Unsure how C name resolution works, so we make sure to version this symbol.
 ///
 /// Return `unsigned char` since it is guaranteed to be `u8` on all platforms.
-unsigned char rust_objc_sys_0_2_try_catch_exception(void (*f)(void *), void *context, id *error) {
+unsigned char rust_objc_sys_0_3_try_catch_exception(void (*f)(void *), void *context, id *error) {
     @try {
         f(context);
         if (error) {