+++ crates/objc2/CHANGELOG.md
@@ -111,6 +111,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   );
 
   ```
+* Updated `ffi` module to `objc-sys v0.3.0`.
 
 ### Fixed
 * Allow empty structs in `declare_class!` macro.